refactored home; added postlist
