+++ frontend/src/components/minimizedPost/minimizedPost.css
@@ -9,6 +9,7 @@
     border-radius: 10px;
     box-shadow: 0 4px 6px rgba(0, 0, 0, 0.2);
     max-width: 500px;
+    margin: 0 auto;
 }
 
 .profile-section {
